Moved item importing to the load screen.
